Tests: Update tests

+++ wordless/wl_dialogs/wl_dialogs.py
@@ -76,8 +76,8 @@ class Wl_Dialog(QDialog):
 
     def move_to_center(self):
         self.move(
-            (self.main.width() - self.width()) / 2,
-            (self.main.height() - self.height()) / 2
+            int((self.main.width() - self.width()) / 2),
+            int((self.main.height() - self.height()) / 2)
         )
 
 class Wl_Dialog_Frameless(Wl_Dialog):
